# openapi/target/classes/openapi.yml
openapi: 3.0.3
info:
  title: API oPROsita
  description: API REST para gestión de grupos, usuarios, contenido y simulacros
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /novedades/profesor:
    get:
      tags:
        - novedades
      operationId: getNovedadesProfesor
      summary: Obtener novedades para profesores
      parameters:
        - in: query
          name: tipo
          required: false
          schema:
            type: string
            enum: [ chat, contenido, todos ]
        - in: query
          name: fechaDesde
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Lista de novedades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Novedad'


    post:
      tags:
        - novedades
      operationId: createNovedadProfesor
      summary: Crear novedad para profesores
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Novedad'
      responses:
        '201':
          description: Novedad creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Novedad'
  /novedades/alumno:
    get:
      tags:
        - novedades
      operationId: getNovedadesAlumno
      summary: Obtener novedades para alumnos
      responses:
        '200':
          description: Lista de novedades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Novedad'

    post:
      tags:
        - novedades
      operationId: createNovedadAlumno
      summary: Crear novedad para alumnos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Novedad'
      responses:
        '201':
          description: Novedad creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Novedad'

    put:
      tags:
        - novedades
      operationId: updateNovedadAlumno
      summary: Actualizar novedad para alumno
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Novedad'
      responses:
        '200':
          description: Novedad actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Novedad'

    delete:
      tags:
        - novedades
      operationId: deleteNovedadesAlumno
      summary: Eliminar todas las novedades para alumno
      responses:
        '204':
          description: Novedades eliminadas correctamente
  /novedades/{id}:
    delete:
      tags:
        - novedades
      operationId: deleteNovedad
      summary: Eliminar una novedad
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Novedad eliminada correctamente
  /noticias:
    get:
      tags:
        - noticias
      operationId: getNoticias
      summary: Obtener noticias
      responses:
        '200':
          description: Lista de noticias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Noticia'

    post:
      tags:
        - noticias
      operationId: createNoticia
      summary: Crear una nueva noticia (PDF + descripción)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                descripcion:
                  type: string
                grupoId:
                  type: integer
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Noticia creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Noticia'
  /noticias/{id}:
    delete:
      tags:
        - noticias
      operationId: deleteNoticia
      summary: Eliminar una noticia
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Noticia eliminada correctamente
...
components:
  schemas:
...
    Novedad:
      type: object
      properties:
        id:
          type: integer
        texto:
          type: string
        tipoDestinatario:
          type: string
          enum: [profesor, alumno]
        fechaCreacion:
          type: string
          format: date-time
    Noticia:
      type: object
      properties:
        id:
          type: integer
        descripcion:
          type: string
        archivoId:
          type: integer
        grupoId:
          type: integer
...